removed unneeded files
