--- zabbix-compnent-with-tag/config.py
@@ -1,12 +1,0 @@
-# Configurations for Zabbix
-zabbix_url = "http://localhost:9999/zabbix"
-zabbix_user = "insight.finder"
-zabbix_password = "2ri2T*%HKxbs"
-
-# Configurations for InsightFinder
-insightfinder_url = "https://stg.insightfinder.com"
-insightfinder_username = "ashvat"
-insightfinder_password = "SynysterAshi@01"
-license_key = "fa28bf0a1a0f6e77c6eea446383b36d0b5a884d2"
-insightfinder_project = "zabbix-metadata-4"
-user_agent = "Mozilla/5.0 (compatible; InsightFinderClient/1.0;)"
