--- zabbix-compnent-with-tag/instance_component.json
@@ -1,12 +1,0 @@
-{
-  "Zabbix server": null,
-  "BL-Pole106-CPE": "CPE-Baicells",
-  "BL-Spare Parts-Pole148-CPE": "CPE-Baicells",
-  "BL-Spare Parts-Cabin168-CPE": "CPE-Baicells",
-  "BL-Spare Parts-Pole166-CPE": "CPE-Baicells",
-  "BL-Spare Parts-Cabin98-CPE": "CPE-Baicells",
-  "BL-Spare Parts-Cabin89-CPE": "CPE-Baicells",
-  "BL-Spare Parts-Cabin94-CPE": "CPE-Baicells",
-  "BL-Spare Parts-Pole3-CPE": "CPE-Baicells",
-  "BL-Spare Parts-CPE": "CPE-Baicells"
-}
